@@ -1,7 +1,7 @@
 {
   "$schema": "./node_modules/ng-packagr/ng-package.schema.json",
   "name": "ngx-input-file",
-  "version": "7.0.4",
+  "version": "8.0.0",
   "description": "Angular Input File Component",
   "license": "MIT",
   "repository": {
@@ -36,6 +36,7 @@
       "entryFile": "public_api.ts"
     }
   },
+  "private": false,
   "peerDependencies": {
     "@angular/animations": "^8.0.0",
     "@angular/cdk": "^8.0.0",
@@ -45,42 +46,40 @@
     "@angular/material": "^8.0.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.800.2",
-    "@angular/animations": "^7.0.1",
-    "@angular/cdk": "^7.0.1",
+    "@angular-devkit/build-angular": "~0.800.0",
+    "@angular/animations": "^8.0.0",
+    "@angular/cdk": "^8.0.0",
     "@angular/cli": "~8.0.2",
-    "@angular/common": "^7.0.1",
-    "@angular/compiler": "^8.0.0",
-    "@angular/compiler-cli": "^8.0.0",
-    "@angular/core": "^7.0.1",
-    "@angular/forms": "^7.0.1",
-    "@angular/http": "^7.2.15",
-    "@angular/language-service": "^8.0.0",
-    "@angular/material": "^7.0.1",
-    "@angular/platform-browser": "^8.0.0",
-    "@angular/platform-browser-dynamic": "^8.0.0",
-    "@angular/router": "^8.0.0",
-    "@types/jasmine": "~3.3.13",
-    "@types/jasminewd2": "~2.0.6",
-    "@types/node": "~12.0.7",
+    "@angular/common": "^8.0.0",
+    "@angular/compiler": "~8.0.0",
+    "@angular/compiler-cli": "~8.0.0",
+    "@angular/core": "^8.0.0",
+    "@angular/forms": "^8.0.0",
+    "@angular/language-service": "~8.0.0",
+    "@angular/material": "^8.0.0",
+    "@angular/platform-browser": "~8.0.0",
+    "@angular/platform-browser-dynamic": "~8.0.0",
+    "@angular/router": "~8.0.0",
+    "@types/jasmine": "~3.3.8",
+    "@types/jasminewd2": "~2.0.3",
+    "@types/node": "~8.9.4",
     "bootstrap": "^4.3.1",
-    "codelyzer": "~5.1.0",
-    "core-js": "^3.1.3",
+    "codelyzer": "^5.0.0",
     "jasmine-core": "~3.4.0",
     "jasmine-spec-reporter": "~4.2.1",
     "karma": "~4.1.0",
     "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "~2.0.5",
+    "karma-coverage-istanbul-reporter": "~2.0.1",
     "karma-jasmine": "~2.0.1",
-    "karma-jasmine-html-reporter": "^1.4.2",
+    "karma-jasmine-html-reporter": "^1.4.0",
     "material-design-icons": "^3.0.1",
     "ng-packagr": "^5.3.0",
-    "protractor": "~5.4.2",
-    "rxjs": "^6.5.2",
-    "ts-node": "~8.2.0",
-    "tsickle": "^0.35.0",
-    "tslint": "~5.17.0",
-    "typescript": "~3.4.1",
-    "zone.js": "^0.9.1"
+    "protractor": "~5.4.0",
+    "rxjs": "~6.4.0",
+    "ts-node": "~7.0.0",
+    "tslib": "^1.9.0",
+    "tslint": "~5.15.0",
+    "typescript": "~3.4.3",
+    "zone.js": "~0.9.1"
   }
 }
