@@ -5,8 +5,8 @@ import { DemoModule } from './demo/demo.module';
 import { environment } from './environments/environment';
 
 if (environment.production) {
-    enableProdMode();
+  enableProdMode();
 }
 
 platformBrowserDynamic().bootstrapModule(DemoModule)
-    .catch(err => console.log(err));
+  .catch(err => console.error(err));
